refactoring updates
