+++ utils/functions/helper.py
@@ -1,11 +1,5 @@
 import streamlit as st
 import json
-from pathlib import Path
-
-LABELS_PATH = Path("conf/base/species_labels.json")
-
-with LABELS_PATH.open("r") as f:
-    SPECIES_LABELS = json.load(f)
 
 @st.cache_data
 def load_help(path: str = "conf/base/help_text.json"):
@@ -23,16 +17,3 @@ def H(key: str) -> str:
     if isinstance(entry, dict) and "help" in entry:
         return entry["help"]
     return ""
-
-def _species_keys(preset: dict):
-    """
-    Accessor for species keys loaded from conf/base/FVSVariant_presets.json.
-    """
-    # any key that starts with tpa_ is treated as a species slider
-    return [k for k in preset.keys() if k.startswith("tpa_")]
-
-def _label_for(key: str) -> str:
-    """
-    Accessor for species labels loaded from conf/base/species_labels.json.
-    """
-    return SPECIES_LABELS.get(key, key.replace("tpa_", "TPA_").upper())
